The library is asynchronous again!
When creating the BufferedBinaryAjax to enable M4A partial loading I made the error of making everything synchronous, probably not the best idea I had in the latest years :P

Each reader is now responsible for loading its own data, instead of returning a range of bytes to preload using readID3Range() the reader has a loadData(data, callback) function that will load the necessary data and invoke the callback when ready.

The ID4 was a bit tricky since there's a need to search the tags through the entire file. A function named loadAtom() takes care of loading only the ranges where the tags are contained.