Clarify the origin of the CC license for Base64
It seems that the website has changed the terms of use and the page for the license doesn't exist anymore.
This change points to the original page that mentions the CC using the wayback machine.
This file is only used for the purpose of displaying the cover art in the examples and is not used by the library to extract the tags.